+++ src/game/collisions/BoundingBox.h
@@ -10,9 +10,9 @@ class BoundingBox {
 
     public:
 
-        int x1, y1, x2, y2;
+        double x1, y1, x2, y2;
 
-        BoundingBox(int x, int y, int width, int height);
+        BoundingBox(double x, double y, double width, double height);
         bool CheckCollision(const Line& line) const;
         bool CheckCollision(const BoundingBox& other) const;
