+++ src/game/collisions/Line.h
@@ -5,7 +5,7 @@
 
 class Line{
     public:
-        int x1, y1, x2, y2;
-        Line(int x1, int y1, int x2, int y2);
+        double x1, y1, x2, y2;
+        Line(double x1, double y1, double x2, double y2);
         bool CheckCollision(const Line& other) const;
 };
